requirements.txt
Django==1.4.6
python-social-auth>=0.1.23
rdflib==3.4.0
rdfextras>=0.4
SPARQLWrapper>=1.5.2
requests>=1.2.3
beautifulsoup4>=4.3.2
